+++ example/packages/base/lib/model/post.dart
@@ -18,11 +18,7 @@ class Post {
   final String title;
   final String body;
 
-  static Iterable<Post> fromJsonArray(String jsonString) {
-    List posts = json.decode(jsonString);
-    return posts.map((item) {
-      final post = Post.fromJson(item as Map<String, dynamic>);
-      return post;
-    });
-  }
+  static Iterable<Post> fromJsonArray(String jsonString) =>
+      (json.decode(jsonString) as List)
+          .map((item) => Post.fromJson(item as Map<String, dynamic>));
 }